TINKERPOP-2132 Fixed threading problem with Client
The full set of changes are best described in changelog and upgrade docs, but basically made it so that a noop message is accepted by the server. The noop is a "do nothing" request and simply returns a success response. The driver sends this noop just after the websocket handshake and forces authentication to occur. In this way authentication is performed out of band with normal requests and the multithreading problem for Client instances goes away.